# docs/swagger.yaml
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      items: {}
      page:
        type: integer
      pageSize:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  user.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      roleId:
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.RoleInfo:
    properties:
      name:
        type: string
      title:
        type: string
    type: object
...
  user.UpdateUserRequest:
    properties:
      email:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      password:
        minLength: 6
        type: string
      roleId:
        type: string
    required:
    - email
    - name
    type: object
  user.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      role:
        $ref: '#/definitions/user.RoleInfo'
      roleId:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@jvair.com
    name: API Support
    url: http://www.jvair.com/support
  description: API para el sistema JVAIR
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JVAIR API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de usuarios
      parameters:
      - description: 'Número de página (por defecto: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamaño de página (por defecto: 10)'
        in: query
        name: pageSize
        type: integer
      - description: Filtrar por nombre
        in: query
        name: name
        type: string
      - description: Filtrar por email
        in: query
        name: email
        type: string
      - description: Filtrar por rol
        in: query
        name: roleId
        type: string
      - description: Filtrar por estado (activo/inactivo)
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "400":
          description: Parámetros de consulta inválidos
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Listar usuarios
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario en el sistema
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Error al decodificar la solicitud o datos inválidos
          schema:
            type: string
        "409":
          description: Email ya está en uso
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Crear usuario
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un usuario (soft delete)
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: ID de usuario inválido
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Eliminar usuario
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Obtiene un usuario por su ID
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: ID de usuario inválido
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener usuario
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Actualiza un usuario existente
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Error al decodificar la solicitud o datos inválidos
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "409":
          description: Email ya está en uso
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Actualizar usuario
      tags:
      - Users
  /api/v1/users/{id}/abilities:
    get:
      consumes:
      - application/json
      description: Obtiene las habilidades asignadas a un usuario
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener habilidades de usuario
      tags:
      - Users
  /api/v1/users/{id}/roles:
    get:
      consumes:
      - application/json
      description: Obtiene los roles asignados a un usuario
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener roles de usuario
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Ingrese el token con el prefijo 'Bearer '
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...